+++ server/src/trade/buy.ts
@@ -1,7 +1,7 @@
 'use strict';
 
-import { BuyResult } from "../types";
 import Broker from "./broker/Broker";
+import { BuyResult } from "./types";
 const broker = new Broker();
 const { TRADE_QTY } = require('./trade.config.json');
