+++ server/src/types.ts
@@ -21,130 +21,3 @@ export type AxiosResponse = {
     config: JSObject,
     request: JSObject,
 };
-
-/***********************************/
-
-export type Source = {
-    name: string,
-    url: string,
-    selectors: {
-        row: string,
-        rating: string,
-        symbol: string,
-        drillDownLink: string,
-    },
-    constants: {
-        BULLISH: string,
-        NEUTRAL: string,
-        BEARISH: string,
-    };
-};
-
-// TODO: Make this generic
-export type BuyOrder = {
-    symbol: string,
-    qty: number,
-    notional?: number,
-    side?: string,
-    type?: string,
-    time_in_force?: string,
-    limit_price?: number,
-    stop_price?: number,
-    trail_price?: number,
-    trail_percent?: number,
-    extended_hours?: boolean,
-    client_order_id?: string,
-    order_class?: string,
-    take_profile?: {
-        limit_price: number,
-    },
-    stop_loss?: {
-        stop_price: number,
-        limit_price: number,
-    },
-};
-
-// TODO: Make this generic
-export type SellOrder = {
-    symbol: string,
-    qty: number,
-    notional?: number,
-    side?: string,
-    type?: string,
-    time_in_force?: string,
-    limit_price?: number,
-    stop_price?: number,
-    trail_price?: number,
-    trail_percent?: number,
-    extended_hours?: boolean,
-    client_order_id?: string,
-    order_class?: string,
-    take_profile?: {
-        limit_price: number;
-    },
-    stop_loss?: {
-        stop_price: number,
-        limit_price: number,
-    },
-};
-
-export type Position = {
-    id: string,
-    symbol: string,
-    qty: number,
-    market_value: number,
-    current_price: number,
-    lastday_price?: number,
-    change_today?: number,
-    broker: string,
-    response: JSObject,
-};
-
-export type BuyResult = {
-    request_id: string,
-    symbol: string,
-    notional?: number,
-    qty: number,
-    security_price: number,
-    broker: string,
-    statusCode: HttpStatusCode,
-    statusText: string,
-    errors?: string[],
-    timestamp?: string,
-    response: any,
-};
-
-export type SellResult = {
-    request_id: string,
-    symbol: string,
-    notional?: number,
-    qty: number,
-    security_price: number,
-    broker: string,
-    statusCode: HttpStatusCode,
-    statusText: string,
-    errors?: string[],
-    timestamp?: string,
-    response: any,
-};
-
-export interface AccountInfo {
-    account_number: string,
-    broker: string,
-    funds: number,
-    response: JSObject,
-};
-
-export type BuyFunction = (buyOrder: BuyOrder) => Promise<BuyResult>;
-export type SellFunction = (sellOrder: SellOrder) => Promise<SellResult>;
-export type GetAccountInfoFunction = () => Promise<AccountInfo>;
-export type GetPositionsFunction = () => Promise<Position[]>;
-export type GetPositionFunction = (symbol: string) => Promise<Position>;
-
-export type BrokerProvider = {
-    buy: BuyFunction,
-    sell: SellFunction,
-    getAccountInfo: GetAccountInfoFunction,
-    getPositions: GetPositionsFunction,
-    getPosition: GetPositionFunction,
-};
