+++ test/unit/server/trade/buy.test.ts
@@ -1,5 +1,3 @@
-const { buySymbols, getBuyableSymbols, buy } = require('../../../../server/src/trade/buy');
-
 describe('Buy module', () => {
     describe('buySymbols()', () => {
         it('DUMMY TEST', () => {
