+++ test/unit/server/trade/sell.test.ts
@@ -1,5 +1,3 @@
-const { sellPositions, getPositions, isPositionSellable, sellPosition } = require('../../../../server/src/trade/sell');
-
 describe('Sell module', () => {
     describe('sellPositions()', () => {
         it('DUMMY TEST', () => {
